feat(hugr-py): Allow defining functions, consts, and aliases inside DFGs
